+++ src/main/java/soma/ghostrunner/domain/member/application/MemberService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import soma.ghostrunner.domain.member.*;
 import soma.ghostrunner.domain.member.domain.Member;
-import soma.ghostrunner.domain.member.MemberBioInfo;
+import soma.ghostrunner.domain.member.domain.MemberBioInfo;
 import soma.ghostrunner.domain.member.MemberNotFoundException;
 import soma.ghostrunner.domain.member.MemberRepository;
 import soma.ghostrunner.domain.member.api.dto.ProfileImageUploadRequest;
