+++ src/main/java/soma/ghostrunner/domain/member/domain/Member.java
@@ -2,7 +2,6 @@ package soma.ghostrunner.domain.member.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import soma.ghostrunner.domain.member.MemberBioInfo;
 import soma.ghostrunner.global.common.BaseTimeEntity;
 import soma.ghostrunner.global.common.document.TestOnly;
